requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
click>=8.1.0
urllib3>=1.26.0
pandas>=2.0.0
openpyxl>=3.1.0
tqdm>=4.65.0
pypdf>=3.0.0
reportlab>=4.0.0
python-docx>=1.1.0
Pillow>=10.0.0
doc2docx>=0.2.0

# WMF/EMF Support Notes:
# - Windows: Works natively with PIL (no additional packages needed)
# - Mac/Linux: Requires ImageMagick + wand package (see README)
#   Install ImageMagick first, then uncomment below:
# wand>=0.6.0
